rag/requirements.txt
langchain
openai
numpy
faiss-cpu
scikit-learn
fastapi
uvicorn
pdfplumber
langchain-community
langchain-openai
